+++ extensions/jotform/CHANGELOG.md
@@ -1,3 +1,6 @@
 # Jotform Changelog
 
 ## [Initial Version] - {PR_MERGE_DATE}
+
+- List Forms in user account
+    - View Submissions in each form